Fix bugs

+++ models.py
@@ -14,7 +14,7 @@ class Generator(nn.Module):
         self.feature_sizes = (self.img_size[0] / 16, self.img_size[1] / 16)
 
         self.latent_to_features = nn.Sequential(
-            nn.Linear(latent_dim, 8 * dim * self.feature_sizes[0] * self.feature_sizes[1]),
+            nn.Linear(latent_dim, int(8 * dim * self.feature_sizes[0] * self.feature_sizes[1])),
             nn.ReLU()
         )
 
@@ -36,7 +36,7 @@ class Generator(nn.Module):
         # Map latent into appropriate size for transposed convolutions
         x = self.latent_to_features(input_data)
         # Reshape
-        x = x.view(-1, 8 * self.dim, self.feature_sizes[0], self.feature_sizes[1])
+        x = x.view(-1, 8 * self.dim, int(self.feature_sizes[0]), int(self.feature_sizes[1]))
         # Return generated image
         return self.features_to_image(x)
 
@@ -69,7 +69,7 @@ class Discriminator(nn.Module):
 
         # 4 convolutions of stride 2, i.e. halving of size everytime
         # So output size will be 8 * (img_size / 2 ^ 4) * (img_size / 2 ^ 4)
-        output_size = 8 * dim * (img_size[0] / 16) * (img_size[1] / 16)
+        output_size = int(8 * dim * (img_size[0] / 16) * (img_size[1] / 16))
         self.features_to_prob = nn.Sequential(
             nn.Linear(output_size, 1),
             nn.Sigmoid()
